feat(object-storage): add search bar on swift
ref: #DATATR-2695
