+++ packages/manager/apps/pci-object-storage/src/pages/object-storage/storage/s3Id/objects/Objects.page.tsx
@@ -1,22 +1,19 @@
 import { useTranslation } from 'react-i18next';
-import { Outlet, useNavigate, useParams } from 'react-router-dom';
+import { Outlet, useParams } from 'react-router-dom';
 import {
-  Button,
-  Label,
   Skeleton,
-  Switch,
   Alert,
   AlertDescription,
+  Switch,
+  Label,
 } from '@datatr-ux/uxlib';
-import { useState, useMemo, useDeferredValue } from 'react';
-import { Plus } from 'lucide-react';
+import { useState } from 'react';
 import { useS3Data } from '../S3.context';
 import { useGetS3Objects } from '@/data/hooks/s3-storage/useGetS3Objects.hook';
-import Guides from '@/components/guides/Guides.component';
 import S3ObjectBrowser from './_components/S3ObjectBrowser.component';
 import { useObjectStorageData } from '@/pages/object-storage/ObjectStorage.context';
 import { useIsLocaleZone } from '@/hooks/useIsLocalZone.hook';
-import SearchBar from './_components/SearchBar.component';
+import ObjectsPageHeader from '@/components/objects-page-header/ObjectsPageHeader.component';
 
 const Objects = () => {
   const { projectId } = useParams();
@@ -25,7 +22,6 @@ const Objects = () => {
   const isLocaleZone = useIsLocaleZone(s3, regions);
   const { t } = useTranslation('pci-object-storage/storages/s3/objects');
   const [withVersion, setWithVersion] = useState(false);
-  const [searchQuery, setSearchQuery] = useState('');
 
   const objectQuery = useGetS3Objects({
     projectId,
@@ -33,17 +29,8 @@ const Objects = () => {
     name: s3.name,
     withVersions: withVersion,
   });
-  const navigate = useNavigate();
 
   const objects = objectQuery.data || [];
-
-  const deferredSearchQuery = useDeferredValue(searchQuery);
-
-  const filteredObjects = useMemo(() => {
-    if (!deferredSearchQuery) return objects;
-    const query = deferredSearchQuery.toLowerCase();
-    return objects.filter((obj) => obj.key?.toLowerCase().includes(query));
-  }, [objects, deferredSearchQuery]);
 
   if (objectQuery.isLoading) return <Objects.Skeleton />;
 
@@ -60,39 +47,18 @@ const Objects = () => {
 
   return (
     <>
-      <div
-        data-testid="containers-guides-container"
-        className="flex justify-between w-full items-center"
-      >
-        <h2>{t('objectTitle')}</h2>
-        <div className="flex flex-wrap justify-end gap-1">
-          <Guides selectors={['allGuides', 'gettingStarted']} />
-        </div>
-      </div>
-
-      <div className="flex justify-between">
-        <Button onClick={() => navigate('./add-object')}>
-          <Plus className="size-6" />
-          {t('addNewObject')}
-        </Button>
+      <ObjectsPageHeader objects={objects}>
         {!isLocaleZone && (
-          <div className="flex items-center space-x-2">
+          <>
             <Switch
               id="versions"
               checked={withVersion}
               onCheckedChange={setWithVersion}
             />
             <Label htmlFor="versions">{t('seeVersionsSwitchLabel')}</Label>
-            <SearchBar
-              searchQuery={searchQuery}
-              onSearchChange={setSearchQuery}
-              deferredSearchQuery={deferredSearchQuery}
-              filteredObjects={filteredObjects}
-              placeholder={t('searchPlaceholder') || 'Search...'}
-            />
-          </div>
+          </>
         )}
-      </div>
+      </ObjectsPageHeader>
       <S3ObjectBrowser
         objects={objects}
         isLocaleZone={isLocaleZone}
